# backend-updated/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: backend

# during dev, should point to your local machine
host: localhost:3001
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
paths:
...
  /notes:
    x-swagger-router-controller: notes
    post:
      description: Adds a new post
      operationId: addNote
      parameters:
        - name: note
          description: Object with the register data
          in: body
          required: true
          schema:
            type: object
            properties:
              noteType:
                type: string
              textNote:
                type: string
              name:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Get all notes
      operationId: findAll
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string